Additional Options
Added option to use AUC instead of p-value as target criterion.

Added option to drop markers at each step, making sure that the dropping
of any marker harms the target criterion. If the dropping of any marker
improves the target criterion, then it is moved to the set of unselected
markers and the search proceeds.